requirements.txt
requests
responses
pyyaml
pytest
flake8
